Fix execvpe() implementation
There is a serious problem with the execvpe implementation that is
included here.  On OpenBSD with a `doas.conf` like

	permit nopass keepenv :wheel cmd foobar

doas will restrict PATH to `/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin`

        envp = prepenv(rule);

        if (rule->cmd) {
                if (setenv("PATH", safepath, 1) == -1)
                        err(1, "failed to set PATH '%s'", safepath);
        }
        execvpe(cmd, argv, envp);

However the execvpe() implementation from `execvpe.c` is flawed and
ignores the previous setenv() call and will use `envp` verbatim as the
new process' environment.

The result is that users can run arbitrary commands as root by simply
setting a new `PATH` with a different `foobar` in it.

Change the execvpe implementation to the one from OpenBSD.

Signed-off-by: Tobias Kortkamp <tobik@FreeBSD.org>